+++ test/optimization/test_partition.py
@@ -18,12 +18,12 @@ import unittest
 from test.optimization import QiskitOptimizationTestCase
 import numpy as np
 from qiskit import BasicAer
+from qiskit.circuit.library import RealAmplitudes
 from qiskit.aqua import aqua_globals, QuantumInstance
 from qiskit.optimization.applications.ising import partition
 from qiskit.optimization.applications.ising.common import read_numbers_from_file, sample_most_likely
 from qiskit.aqua.algorithms import NumPyMinimumEigensolver, VQE
 from qiskit.aqua.components.optimizers import SPSA
-from qiskit.aqua.components.variational_forms import RY
 
 
 class TestSetPacking(QiskitOptimizationTestCase):
@@ -48,13 +48,13 @@ class TestSetPacking(QiskitOptimizationTestCase):
         """ Partition VQE test """
         aqua_globals.random_seed = 100
         result = VQE(self.qubit_op,
-                     RY(self.qubit_op.num_qubits, depth=5, entanglement='linear'),
+                     RealAmplitudes(reps=5, entanglement='linear'),
                      SPSA(max_trials=200),
                      max_evals_grouped=2).run(
                          QuantumInstance(BasicAer.get_backend('qasm_simulator'),
                                          seed_simulator=aqua_globals.random_seed,
                                          seed_transpiler=aqua_globals.random_seed))
-        x = sample_most_likely(result['eigvecs'][0])
+        x = sample_most_likely(result.eigenstate)
         self.assertNotEqual(x[0], x[1])
         self.assertNotEqual(x[2], x[1])  # hardcoded oracle